using size_t

+++ tasks/levonychev_i_radix_batcher_sort/seq/include/ops_seq.hpp
@@ -17,7 +17,7 @@ class LevonychevIRadixBatcherSortSEQ : public BaseTask {
   bool PreProcessingImpl() override;
   bool RunImpl() override;
   bool PostProcessingImpl() override;
-  static void CountingSort(InType &arr, int byte_index);
+  static void CountingSort(InType &arr, size_t byte_index);
 };
 
 }  // namespace levonychev_i_radix_batcher_sort
