+++ tasks/levonychev_i_radix_batcher_sort/seq/src/ops_seq.cpp
@@ -13,14 +13,14 @@ LevonychevIRadixBatcherSortSEQ::LevonychevIRadixBatcherSortSEQ(const InType &in)
   GetInput() = in;
 }
 
-void LevonychevIRadixBatcherSortSEQ::CountingSort(InType &arr, int byte_index) {
-  const int byte = 256;
+void LevonychevIRadixBatcherSortSEQ::CountingSort(InType &arr, size_t byte_index) {
+  const size_t byte = 256;
   std::vector<int> count(byte, 0);
   OutType result(arr.size());
 
-  bool is_last_byte = (byte_index == (sizeof(int) - 1));
+  bool is_last_byte = (byte_index == (sizeof(int) - 1ULL));
   for (auto number : arr) {
-    int value_of_byte = (number >> (byte_index * 8)) & 0xFF;
+    int value_of_byte = (number >> (byte_index * 8ULL)) & 0xFF;
 
     if (is_last_byte) {
       value_of_byte ^= 0x80;
@@ -29,12 +29,12 @@ void LevonychevIRadixBatcherSortSEQ::CountingSort(InType &arr, int byte_index) {
     ++count[value_of_byte];
   }
 
-  for (int i = 1; i < byte; ++i) {
+  for (size_t i = 1ULL; i < byte; ++i) {
     count[i] += count[i - 1];
   }
 
   for (int i = arr.size() - 1; i >= 0; --i) {
-    int value_of_byte = (arr[i] >> (byte_index * 8)) & 0xFF;
+    int value_of_byte = (arr[i] >> (byte_index * 8ULL)) & 0xFF;
 
     if (is_last_byte) {
       value_of_byte ^= 0x80;
@@ -61,7 +61,7 @@ bool LevonychevIRadixBatcherSortSEQ::RunImpl() {
   }
   std::cout << std::endl;
 
-  for (int i = 0; i < sizeof(int); ++i) {
+  for (size_t i = 0; i < sizeof(int); ++i) {
     CountingSort(GetOutput(), i);
   }
